Encapsulate base64-decoding of Sinch Application Secret
Purpose with this change is to highlight that the Sinch Application
Secret is orginally aquired in the form of a base64-encoded string,
and it has to be base64-decoded before used as signing key data.

NOTE: This is most likely not how it would be implemented in
production code (as it would mean the base64-decoding would have to be
done over and over again), but, these are examples with a focus on
readability and clarity.