Support HMS Application ID as JWT 'sub' claim
- HMS Application ID is passed as 'sub' claim.
- Sinch Application Key is now passed as JWT header parameter (instead of as
'sub' claim), and as custom claim 'sinch:rtc:application_key'.

NOTE: The reason that the Sinch Application Key is passed both as a
JWT header parameter and as a signed claim is that it allows an
implementation to truly separate the use of the Application Key for
looking up a signing key (JWT header param), and for any use in
processing. I.e. once the JWT has been validated, all data necessary
to process the request should be self-contained in the payload and
signed claims.

(These are adaptions to changes outlined in
https://github.com/sinch/docs/pull/180 and https://github.com/sinch/docs/pull/181)